Fixing the Jump mechanics
Due to unmatching height between the Skeletal Mesh's (due to corresponding animations had some height) and the Capsule Comp when jumping. I've adjusted the height in the Jumps animation so that it matches with the Capsule Comp AMAP. Right now, it's close enough. Definitely can extend the tweaking to make it look better